Changing code to avoid recursion functionality error in python.
